back/requirements/prod.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
passlib
bcrypt
python-jose
email_validator
requests
httpx
google-cloud-storage
google-auth
google-auth-httplib2
googleapis-common-protos
protobuf
proto-plus
sse-starlette
starlette
greenlet
cryptography
pyasn1
pyasn1_modules
rsa
cachetools
typing_extensions
google-adk==1.3.0
python-slugify
alembic
redis
